Rename test folder to tests.
